test(tests): Move location of test files and add graphql tests
